refactor(swc/common): bubble up sourcemapperext
