+++ crates/swc_common/Cargo.toml
@@ -33,6 +33,7 @@ anyhow = {version = "1.0.45", optional = true}
 arbitrary = {version = "1", optional = true, features = ["derive"]}
 ast_node = {version = "0.7.5", path = "../ast_node"}
 atty = {version = "0.2", optional = true}
+bitflags = "1"
 better_scoped_tls = {version = "0.1.0", path = "../better_scoped_tls"}
 cfg-if = "1.0.0"
 debug_unreachable = "0.1.1"
