+++ crates/swc_common/src/lib.rs
@@ -34,6 +34,7 @@
 #![deny(clippy::all)]
 #![deny(unused)]
 #![cfg_attr(docsrs, feature(doc_cfg))]
+#![recursion_limit = "1024"]
 
 use std::fmt::Debug;
 
@@ -46,6 +47,7 @@ pub use swc_visit::chain;
 pub use self::{
     eq::{EqIgnoreSpan, TypeEq},
     errors::{SourceMapper, SourceMapperDyn},
+    list::ListFormat,
     pos::{
         hygiene, BytePos, CharPos, FileName, Globals, Loc, LocWithOpt, Mark, MultiSpan, SourceFile,
         SourceFileAndBytePos, SourceFileAndLine, Span, SpanLinesError, Spanned, SyntaxContext,
@@ -62,6 +64,8 @@ pub trait AstNode: Debug + PartialEq + Clone + Spanned + Serialize {
     const TYPE: &'static str;
 }
 
+#[macro_use]
+pub mod list;
 pub mod collections;
 pub mod comments;
 mod eq;
