+++ crates/swc_common/src/list.rs
@@ -1,6 +1,8 @@
+#![allow(unused)]
 #![allow(non_upper_case_globals)]
 use bitflags::bitflags;
-use swc_common::add_bitflags;
+
+use crate::add_bitflags;
 
 bitflags! {
     pub struct ListFormat: u32 {
