+++ crates/swc_ecma_codegen/Cargo.toml
@@ -13,7 +13,6 @@ version = "0.108.6"
 bench = false
 
 [dependencies]
-bitflags = "1"
 memchr = "2.4.1"
 num-bigint = { version = "0.4", features = ["serde"] }
 once_cell = "1.10.0"
