+++ crates/swc_ecma_codegen_macros/src/lib.rs
@@ -63,7 +63,7 @@ fn (&mut self, node: Node) -> Result;
                             fn emit_with<W, S: swc_common::SourceMapper>(&self, e: &mut crate::Emitter<'_, W, S>) -> Result
                             where
                                 W: crate::text_writer::WriteJs,
-                                S: crate::util::SourceMapperExt
+                                S: swc_common::source_map::SourceMapperExt
                             {
                                 e.mtd_name(self)
                             }
